Move conversion to a separate module to help rust-analyzer.
